refactor(indexer): move IndexVerification to index-context, remove unsafe type casts (R1)
Move IndexVerification type definition from indexer.ts to its proper home in
index-context.ts alongside IndexRunContext. Replace unsafe type casts (ctx as
IndexRunContext & { _verification; _totalEntries }) with direct property access,
establishing a clean semantic boundary and improving code safety.
